requirements.txt

numpy
pandas
matplotlib
pillow
pytesseract
